+++ agent/discovery/src/config.rs
@@ -1,0 +1,14 @@
+/// Agent discovery configuration options.
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct Config {
+    pub files: Vec<String>,
+}
+
+impl Default for Config {
+    fn default() -> Config {
+        Config {
+            files: Vec::new(),
+        }
+    }
+}