+++ agent/discovery/src/errors.rs
@@ -1,0 +1,6 @@
+error_chain! {
+    foreign_links {
+        IoError(::std::io::Error);
+        YamlDecode(::serde_yaml::Error);
+    }
+}